EDGRTAC-29 Rtac batch support
