+++ ramls/batch-holdings.xsd
@@ -1,26 +1,35 @@
 <?xml version="1.0"?>
 <xs:schema xmlns:xs="http://www.w3.org/2001/XMLSchema">
 
-<xs:element name="holdings">
-  <xs:complexType>
-    <xs:sequence>
-      <xs:element ref="holding" maxOccurs="unbounded" minOccurs="0"/>
-    </xs:sequence>
-  </xs:complexType>
-</xs:element>
+  <xs:element name="instances">
+    <xs:complexType>
+      <xs:sequence>
+        <xs:element ref="holdings"/>
+      </xs:sequence>
+    </xs:complexType>
+  </xs:element>
 
-<xs:element name="holding">
-  <xs:complexType>
-    <xs:sequence>
-      <xs:element name="id" type="xs:string" minOccurs="1"/>
-      <xs:element name="callNumber" type="xs:string"/>
-      <xs:element name="location" type="xs:string"/>
-      <xs:element name="status" type="xs:string"/>
-      <xs:element name="dueDate" type="xs:string"/>
-      <xs:element name="tempLocation" type="xs:string" minOccurs="0"/>
-      <xs:element name="volume" type="xs:string" minOccurs="0"/>
-    </xs:sequence>
-  </xs:complexType>
-</xs:element>
+  <xs:element name="holdings">
+    <xs:complexType>
+      <xs:sequence>
+        <xs:element name="instanceId" type="xs:string"/>
+        <xs:element ref="holding" maxOccurs="unbounded" minOccurs="0"/>
+      </xs:sequence>
+    </xs:complexType>
+  </xs:element>
+
+  <xs:element name="holding">
+    <xs:complexType>
+      <xs:sequence>
+        <xs:element name="id" type="xs:string" minOccurs="1"/>
+        <xs:element name="callNumber" type="xs:string"/>
+        <xs:element name="location" type="xs:string"/>
+        <xs:element name="status" type="xs:string"/>
+        <xs:element name="dueDate" type="xs:string"/>
+        <xs:element name="tempLocation" type="xs:string" minOccurs="0"/>
+        <xs:element name="volume" type="xs:string" minOccurs="0"/>
+      </xs:sequence>
+    </xs:complexType>
+  </xs:element>
 
 </xs:schema>
